backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
sqlalchemy
psycopg[binary]
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
email-validator